requirements.txt
streamlit
pandas
numpy
bokeh
tensorflow
scikit-learn
vertexai
google-auth
google-cloud-aiplatform
numpy
faiss-cpu
PyMuPDF
PyPDF2
